VxWorks fixes
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@1273 42547f41-83fc-0310-9aef-9668e019ab9e

+++ motor/marea_detector.c
@@ -452,7 +452,7 @@ motor_area_detector_fn( int argc, char *argv[] )
 	"    situation is to restart the MX server.  If you ever actually\n"
 	"    see this error message, please tell Bill Lavender about it.\n",
 			cname, starting_total_num_frames, num_frames,
-			LONG_MAX, LONG_MAX );
+			(long) LONG_MAX, (long) LONG_MAX );
 
 			return FAILURE;
 		}
